// src/components/containers/spaces/core-types.ts
export enum Type {
  ViewPort = 'viewport',
  Fixed = 'fixed',
  Fill = 'fill',
  Positioned = 'positioned',
  Anchored = 'anchored',
  Custom = 'custom',
}

export enum AnchorType {
  Left = 'anchor-left',
  Right = 'anchor-right',
  Top = 'anchor-top',
  Bottom = 'anchor-bottom',
}

export enum Orientation {
  Horizontal,
  Vertical,
}

export type SizeUnit = number | string | undefined;

export enum ResizeType {
  All = 'resize-all',
  Left = 'resize-left',
  Right = 'resize-right',
  Top = 'resize-top',
  Bottom = 'resize-bottom',
  TopLeft = 'resize-topleft',
  TopRight = 'resize-topright',
  BottomLeft = 'resize-bottomleft',
  BottomRight = 'resize-bottomright',
}

export enum ResizeHandlePlacement {
  OverlayInside = 'overlay-inside',
  Inside = 'inside',
  OverlayBoundary = 'overlay-boundary',
}